chore(suite): onboarding tests move to package
