+++ suite/onboarding/src/__fixtures__/onboardingActions.ts
